fix(ci): python-coverage-shards — split oversized operator-hint test modules under 1500-line limit
test_pr_monitor_operator_hints_part_002.py (1753) and _part_003.py (1589)
both grew past the 1500-line first-party maintainability cap enforced by
test_core_decomposition_maintainability.py, failing the coverage shard.

Move the pause/drop-restart cases into a new _part_005.py and the
validation-fix-pass cases into _part_006.py (carrying only the fixtures and
helpers each slice uses), dropping now-unused imports. No behavior change;
all moved and retained tests still pass.
